Move ExitCodeDecoder to internal; add basic test
