+++ gshell-core/src/main/java/com/planet57/gshell/internal/ExitCodeDecoder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.planet57.gshell;
+package com.planet57.gshell.internal;
 
 import com.planet57.gossip.Log;
 import com.planet57.gshell.command.CommandAction;
